Database error handling added. Log statement improved

+++ src/movie/movie.service.ts
@@ -5,6 +5,8 @@ import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { MovieNotFoundException } from './exceptions/movie.not.found.exception';
 import { Movie } from './movie.interface';
+import { InvaldMovieIdException } from './exceptions/invalid.movie.id.exception';
+import { DATABASE_EXCEPTION, DatabaseException } from 'src/exceptions/database.exception';
 
 @Injectable()
 export class MovieService {
@@ -14,7 +16,10 @@ export class MovieService {
 
 
     async getMovie(id: string): Promise<Movie> {
-        this.loggerService.log(`findById called with id ${id}`);
+        this.loggerService.log(`Get movied called with id ${id}`);
+        if (id === undefined || id === null) {
+            throw new InvaldMovieIdException('Movie id cannot be null or undefined');
+        }
         try {
             const movie = await this.movieModel.findById(id).exec();
             if (!movie) {
@@ -23,7 +28,7 @@ export class MovieService {
             return movie;
         } catch (error) {
             this.loggerService.error(`Error finding movie: ${error.message}`, error.stack);
-            throw error;
+            throw new DatabaseException(`Error while fetching movie: ${id}`, error);
         }
     }
 
@@ -35,14 +40,17 @@ export class MovieService {
             return createdMovie;
         } catch (error) {
             this.loggerService.error(`Error creating movie: ${error.message}`, error.stack);
-            throw error;
+            throw new DatabaseException(`Error while creating movie with title ${createMovieDto.title}`, error);
         }
 
     }
 
-    async updateMovie(id: string, updateMovieDto: UpdateMovieDto) : Promise<Movie> {
+    async updateMovie(id: string, updateMovieDto: UpdateMovieDto): Promise<Movie> {
         // logic to update a movie by id
         this.loggerService.log(`update movie called with id ${id} and title ${updateMovieDto.title}`);
+        if (id === undefined || id === null) {
+            throw new InvaldMovieIdException('Movie id cannot be null or undefined');
+        }
         try {
             const updatedMovie = await this.movieModel.findByIdAndUpdate(id, updateMovieDto, { new: true }).exec();
             if (!updatedMovie) {
@@ -51,12 +59,16 @@ export class MovieService {
             return updatedMovie;
         } catch (error) {
             this.loggerService.error(`Error updating movie: ${error.message}`, error.stack);
-            throw error;
+            throw new DatabaseException(`Error while updating movie with id ${id}`, error);
+
         }
     }
 
     async deleteMovie(movieId: string): Promise<void> {
         // logic to delete a movie by id
+        if (movieId === undefined || movieId === null) {
+            throw new InvaldMovieIdException('Movie id cannot be null or undefined');
+        }
         try {
             const result = await this.movieModel.findByIdAndDelete(movieId);
             if (!result) {
@@ -64,7 +76,7 @@ export class MovieService {
             }
         } catch (error) {
             this.loggerService.error(`Error deleting movie: ${error.message}`, error.stack);
-            throw error;
+            throw new DatabaseException(`Error while deleting movie with id ${movieId}`, error);
         }
     }
 }
