+++ src/tvshow/tvshow.service.ts
@@ -4,31 +4,37 @@ import { Model } from 'mongoose';
 import { CreateTVShowDto, UpdateTVShowDto } from './tvshow.dto';
 import { TVShow } from './tvshow.interface';
 import { LoggerService } from '../logger/logger.service'; // Add this import
-import {  TVShowNotFoundException } from './exceptions/tvshow.not.found.exception';
+import { TVShowNotFoundException } from './exceptions/tvshow.not.found.exception';
+import { DatabaseException } from 'src/exceptions/database.exception';
+import { InvalidTVShowIdException } from './exceptions/invalid.tvshow.id.exception';
 
 @Injectable()
 export class TVShowService {
   constructor(
     @InjectModel('TVShow') private readonly tvShowModel: Model<TVShow>,
     private readonly logger: LoggerService, // Inject LoggerService
-  ) {}
+  ) { }
 
   async createTVShow(createTVShowDto: CreateTVShowDto): Promise<TVShow> {
-  
-    this.logger.log('Creating a new TV show');
+
+    this.logger.log('Creating a new TV show with title: ' + createTVShowDto.title);
     try {
       const createdTVShow = new this.tvShowModel(createTVShowDto);
       return await createdTVShow.save();
     } catch (error) {
       this.logger.error(`Error creating TV show: ${error.message}`, error.stack);
-      throw error;
+      throw new DatabaseException(`Error creating TV show with Ttile: ${createTVShowDto.title}`, error);
     }
   }
- 
+
 
 
   async getTVShow(tvShowId: string): Promise<TVShow> {
     this.logger.log(`Fetching TV show with id: ${tvShowId}`);
+
+    if (tvShowId === undefined || tvShowId === null) {
+      throw new InvalidTVShowIdException('Movie id cannot be null or undefined');
+    }
     try {
       const tvShow = await this.tvShowModel.findById(tvShowId).exec();
       if (!tvShow) {
@@ -37,12 +43,15 @@ export class TVShowService {
       return tvShow;
     } catch (error) {
       this.logger.error(`Error fetching TV show: ${error.message}`, error.stack);
-      throw error;
+      throw new DatabaseException(`Error while fetching TV show: ${tvShowId}`, error);
     }
   }
 
   async updateTVShow(tvShowId: string, updateTVShowDto: UpdateTVShowDto): Promise<TVShow> {
     this.logger.log(`Updating TV show with id: ${tvShowId}`);
+    if (tvShowId === undefined || tvShowId === null) {
+      throw new InvalidTVShowIdException('Movie id cannot be null or undefined');
+    }
     try {
       const updatedTVShow = await this.tvShowModel.findByIdAndUpdate(tvShowId, updateTVShowDto, { new: true }).exec();
       if (!updatedTVShow) {
@@ -51,12 +60,15 @@ export class TVShowService {
       return updatedTVShow;
     } catch (error) {
       this.logger.error(`Error updating TV show: ${error.message}`, error.stack);
-      throw error;
+      throw new DatabaseException(`Error while updating TV show: ${tvShowId}`, error);
     }
   }
 
   async removeTVShow(tvShowId: string): Promise<void> {
     this.logger.log(`Removing TV show with id: ${tvShowId}`);
+    if (tvShowId === undefined || tvShowId === null) {
+      throw new InvalidTVShowIdException('Movie id cannot be null or undefined');
+    }
     try {
       const result = await this.tvShowModel.findByIdAndDelete(tvShowId);
       if (!result) {
@@ -64,7 +76,7 @@ export class TVShowService {
       }
     } catch (error) {
       this.logger.error(`Error removing TV show: ${error.message}`, error.stack);
-      throw error;
+      throw new DatabaseException(`Error while removing TV show: ${tvShowId}`, error);
     }
   }
